Commented out double swap, fixed priority queue bug
Need to fix priority queue comparison bug, sometimes the elo differences are not ordered correctly.

I think, and I THINK, the balancer for the single swap works correctly. It takes in the elo of whatever value is from team1 and finds the closest value to it (thus finding the smallest absolute value) and switches with it. CHECK THE ELO SWAP MATH, it might be wrong idk.