Created a band-aid fix on loading api's, fixed decimal point issue
The elo's now have decimal places up to 2 now. Only issue is that the List team's page does not call the api using useSWRImmutable. If we just use useSWR, the api is called, but it is also called everytime the person "focuses" on the page. Need to find better result